高版本depot_tools使用最新的

+++ android_armv7.sh
@@ -43,9 +43,11 @@ sudo ldconfig
 cd ~
 echo "=====[ Getting Depot Tools ]====="	
 git clone -q https://chromium.googlesource.com/chromium/tools/depot_tools.git
-cd depot_tools
-git reset --hard 8d16d4a
-cd ..
+if [[ "$VERSION" != "10.6.194" && "$VERSION" != "11.8.172" && "$VERSION" != "13.6.97" ]]; then
+    cd depot_tools
+    git reset --hard 8d16d4a
+    cd ..
+fi
 export DEPOT_TOOLS_UPDATE=0
 export PATH=$(pwd)/depot_tools:$PATH
 gclient
